+++ vsx/src/test/extension.test.ts
@@ -1,15 +1,19 @@
 import * as assert from 'assert';
-
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
 import * as vscode from 'vscode';
-// import * as myExtension from '../../extension';
 
-suite('Extension Test Suite', () => {
-	vscode.window.showInformationMessage('Start all tests.');
+suite('HCL Schema extension', () => {
+	test('activates and registers its commands', async () => {
+		const ext = vscode.extensions.getExtension('avestura.hcl-schema');
+		assert.ok(ext, 'extension should be discoverable');
+		await ext!.activate();
 
-	test('Sample test', () => {
-		assert.strictEqual(-1, [1, 2, 3].indexOf(5));
-		assert.strictEqual(-1, [1, 2, 3].indexOf(0));
+		const commands = await vscode.commands.getCommands(true);
+		for (const id of [
+			'hcl-schema.validateActive',
+			'hcl-schema.restart',
+			'hcl-schema.showOutput',
+		]) {
+			assert.ok(commands.includes(id), `command ${id} should be registered`);
+		}
 	});
 });